refactor: move types to types
